# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "host_id": {
            "title": "Host Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "progress": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/JobStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "task": {
            "readOnly": true,
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "host_id",
          "status",
          "created_at",
          "task"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobError": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "title": "JobError",
        "type": "object"
      },
      "JobKind": {
        "enum": [
          "inventory",
          "preflight",
          "os_network",
          "install",
          "os_capture"
        ],
        "title": "JobKind",
        "type": "string"
      },
...
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobStep": {
        "description": "One named phase of a job (e.g. \"Build installer ISO\"), so progress and failures are legible.",
        "properties": {
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStepStatus",
            "default": "pending"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "key",
          "title"
        ],
        "title": "JobStep",
        "type": "object"
      },
      "JobStepStatus": {
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed",
          "skipped",
          "cancelled"
        ],
        "title": "JobStepStatus",
        "type": "string"
      },
...
      "ProfileSummary": {
        "properties": {
          "default_variant": {
            "title": "Default Variant",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/Variant"
            },
            "title": "Variants",
            "type": "array"
          }
        },
        "required": [
          "id",
          "title",
          "source",
          "default_variant",
          "variants"
        ],
        "title": "ProfileSummary",
        "type": "object"
      },
...
      "Stage": {
        "enum": [
          "hardware",
          "os",
          "readiness",
          "prep",
          "holodeck"
        ],
        "title": "Stage",
        "type": "string"
      },
...
      "TaskInfo": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "destructive": {
            "title": "Destructive",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "optional": {
            "title": "Optional",
            "type": "boolean"
          },
          "produces": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Produces"
          },
          "requires": {
            "items": {
              "type": "string"
            },
            "title": "Requires",
            "type": "array"
          },
          "stage": {
            "$ref": "#/components/schemas/Stage"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "stage",
          "description",
          "produces",
          "requires",
          "optional",
          "destructive",
          "available"
        ],
        "title": "TaskInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Variant": {
        "properties": {
          "cores": {
            "title": "Cores",
            "type": "integer"
          },
          "disk_tb": {
            "title": "Disk Tb",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "memory_gb": {
            "title": "Memory Gb",
            "type": "number"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "cores",
          "memory_gb",
          "disk_tb"
        ],
        "title": "Variant",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "API token (see `groundzero token show`)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Bare metal \u2192 ESXi \u2192 VMware Holodeck, driven through one REST API.",
    "title": "GroundZero API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "host_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Host Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Jobs Api V1 Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/diagnostics": {
      "get": {
        "description": "A redacted debugging bundle: the job and its steps, the host and BMC identity, and every\nRedfish/ESXi exchange (method, path, status, timing, truncated body). Passwords, tokens and session\nsecrets are removed. Attach it to a bug report when a BMC behaves unexpectedly.",
        "operationId": "job_diagnostics_api_v1_jobs__job_id__diagnostics_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Job Diagnostics Api V1 Jobs  Job Id  Diagnostics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Job Diagnostics",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "operationId": "job_events_api_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Server-sent JobEvent stream"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/profiles": {
      "get": {
        "operationId": "list_profiles_api_v1_profiles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileSummary"
                  },
                  "title": "Response List Profiles Api V1 Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Profiles",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/v1/tasks": {
      "get": {
        "description": "The task catalog: what each task needs (``requires``) and makes (``produces``), in pipeline order.",
        "operationId": "list_tasks_api_v1_tasks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskInfo"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tasks",
        "tags": [
          "meta"
        ]
      }
    },
...
  }
}